separate aoi actions and reducers in a different file and migrate createAOI to use SAGA
