+++ src/components/filters/filters_header.js
@@ -1,12 +1,8 @@
 // @flow
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { Map } from 'immutable';
 
 import { Button } from '../button';
-import { Dropdown } from '../dropdown';
-import { AOIManager } from './aoi_manager';
-import { isProd } from '../../config';
 
 class SaveAOI extends React.PureComponent {
   constructor(props) {
