+++ src/store/filters_reducer.js
@@ -1,7 +1,7 @@
 /* @flow */
 import { List, Map, fromJS } from 'immutable';
 
-import { FILTERS, AOI } from './filters_actions';
+import { FILTERS } from './filters_actions';
 
 export type filtersReducerType = Map<
   'loading' | 'error' | 'filters' | 'aoi',
@@ -22,18 +22,6 @@ export function filtersReducer(
   switch (action.type) {
     case FILTERS.set: {
       return state.set('filters', action.filters).set('loading', false);
-    }
-    case AOI.loading: {
-      return state.set('loading', true);
-    }
-    case AOI.clear: {
-      return state.set('aoi', Map()).set('loading', false).set('error', null);
-    }
-    case AOI.fetched: {
-      return state
-        .set('loading', false)
-        .set('error', null)
-        .set('aoi', action.aoi);
     }
     default:
       return state;
