+++ src/store/index.js
@@ -13,12 +13,14 @@ import { changesetsPageReducer } from './changesets_page_reducer';
 import { changesetReducer } from './changeset_reducer';
 import { modalReducer } from './modal_reducer';
 import { filtersReducer } from './filters_reducer';
+import { aoiReducer } from './aoi_reducer';
 
 import type { ModalType } from './modal_reducer';
 import type { ChangesetsPageType } from './changesets_page_reducer';
 import type { ChangesetType } from './changeset_reducer';
 import type { AuthType } from './auth_reducer';
 import type { filtersReducerType } from './filters_reducer';
+import type { aoiReducerType } from './aoi_reducer';
 // Sagas
 import sagas from './sagas';
 
@@ -26,6 +28,7 @@ export type RootStateType = {
   auth: AuthType,
   changesetsPage: ChangesetsPageType,
   filters: filtersReducerType,
+  aoi: aoiReducerType,
   changeset: ChangesetType,
   modal: ModalType,
   routing: Object
@@ -36,6 +39,7 @@ const reducers = combineReducers({
   changesetsPage: changesetsPageReducer,
   changeset: changesetReducer,
   filters: filtersReducer,
+  aoi: aoiReducer,
   routing: routerReducer,
   auth: authReducer,
   modal: modalReducer
