+++ src/store/sagas.js
@@ -3,7 +3,9 @@ import { watchChangesetsPage } from './changesets_page_actions';
 import { watchChangeset, watchModifyChangeset } from './changeset_actions';
 import { watchAuth } from './auth_actions';
 import { watchModal } from './modal_actions';
-import { watchFilters, watchAOI } from './filters_actions';
+import { watchFilters } from './filters_actions';
+import { watchAOI } from './aoi_actions';
+
 export default function* rootSaga() {
   yield all([
     watchChangesetsPage(),
